Schematic and Layout Board for module edition (v1)
This is the version of the device that utilizes modules for the sensors instead of raw ICs. The next iteration will utilize the raw ICs in the PCB design. This PCB was designed in Eagle. All the parts for this edition will be purchased on Amazon and the PCB will be manufactured through JLCPCB and soldered using a soldering iron and header pins after it is tested on a breadboard.